# CraftyClown/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This Api are under cc alias",
    "version": "1.0.0",
    "title": "EAN",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "softwaremail30@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "bunksheet.ml",
  "basePath": "/cc",
  "tags": [
    {
      "name": "Admin",
      "description": "All About Admin"
    }
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/adminCredential": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Check Admin username and Password ",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Username And Password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Cred"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Invalid Credentials"
          }
        }
      }
    },
    "/masterPassword": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Check Master Password",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Username And Password",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CredMaster"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Invalid Credentials"
          }
        }
      }
    },
    "/getEmail": {
      "post": {
        "tags": [
          "Admin"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "id",
            "description": "ID given",
            "required": true,
            "type": "string"
          }
        ],
        "summary": "Get Email From ID",
        "description": "Get Email From ID",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
    "Cred": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "CredMaster": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string"
        }
      }
    }
  }
}